+++ src/Audit.SignalR/ConfigurationApi/AuditHubConfigurator.cs
@@ -3,19 +3,19 @@ using Audit.Core;
 
 namespace Audit.SignalR.Configuration
 {
-    internal class PipelineBuilder : IPipelineBuilder
+    internal class AuditHubConfigurator : IAuditHubConfigurator
     {
         internal string _eventType;
         internal bool _includeHeaders;
         internal bool _includeQueryString;
         internal bool _auditDisabled;
+        internal IAuditHubFilterConfigurator _filters;
         internal EventCreationPolicy? _creationPolicy;
         internal AuditDataProvider _dataProvider;
-        internal IPipelineBuilderFilters _filters;
 
-        public IPipelineBuilder Filters(Action<IPipelineBuilderFilters> config)
+        public IAuditHubConfigurator Filters(Action<IAuditHubFilterConfigurator> config)
         {
-            _filters = new PipelineBuilderFilters();
+            _filters = new AuditHubFilterConfigurator();
             config.Invoke(_filters);
             return this;
         }
@@ -25,31 +25,31 @@ namespace Audit.SignalR.Configuration
             _auditDisabled = true;
         }
 
-        public IPipelineBuilder EventType(string eventType)
+        public IAuditHubConfigurator EventType(string eventType)
         {
             _eventType = eventType;
             return this;
         }
 
-        public IPipelineBuilder IncludeHeaders(bool include = true)
+        public IAuditHubConfigurator IncludeHeaders(bool include = true)
         {
             _includeHeaders = include;
             return this;
         }
 
-        public IPipelineBuilder IncludeQueryString(bool include = true)
+        public IAuditHubConfigurator IncludeQueryString(bool include = true)
         {
             _includeQueryString = include;
             return this;
         }
 
-        public IPipelineBuilder WithCreationPolicy(EventCreationPolicy? policy)
+        public IAuditHubConfigurator WithCreationPolicy(EventCreationPolicy? policy)
         {
             _creationPolicy = policy;
             return this;
         }
 
-        public IPipelineBuilder WithDataProvider(AuditDataProvider provider)
+        public IAuditHubConfigurator WithDataProvider(AuditDataProvider provider)
         {
             _dataProvider = provider;
             return this;