+++ src/Audit.SignalR/SignalrEventDisconnect.cs
@@ -1,7 +1,13 @@
 ﻿using System.Collections.Generic;
+#if ASP_NET
 using Microsoft.AspNet.SignalR.Hubs;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+#else
+using Microsoft.AspNetCore.SignalR;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+#endif
 
 namespace Audit.SignalR
 {
@@ -10,22 +16,28 @@ namespace Audit.SignalR
     /// </summary>
     public class SignalrEventDisconnect : SignalrEventBase
     {
-        [JsonProperty(Order = 5)]
+#if ASP_NET
         [JsonConverter(typeof(StringEnumConverter))]
+#else
+        [JsonConverter(typeof(JsonStringEnumConverter))]
+#endif
         public override SignalrEventType EventType => SignalrEventType.Disconnect;
-        [JsonProperty(Order = 10, DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public bool StopCalled { get; set; }
-        [JsonProperty(Order = 15)]
-        public string ConnectionId { get; set; }
-        [JsonProperty(Order = 20, NullValueHandling = NullValueHandling.Ignore)]
-        public IDictionary<string, string> Headers { get; set; }
-        [JsonProperty(Order = 30, NullValueHandling = NullValueHandling.Ignore)]
-        public IDictionary<string, string> QueryString { get; set; }
-        [JsonProperty(Order = 40, NullValueHandling = NullValueHandling.Ignore)]
+
+#if ASP_NET
         public string LocalPath { get; set; }
-        [JsonProperty(Order = 50, NullValueHandling = NullValueHandling.Ignore)]
+        public bool StopCalled { get; set; }
+#endif
+
+        public string ConnectionId { get; set; }
+        public IDictionary<string, string> Headers { get; set; }
+        public IDictionary<string, string> QueryString { get; set; }
         public string IdentityName { get; set; }
+
+#if ASP_NET
         [JsonIgnore]
         public IHub HubReference { get; set; }
+#else
+        public string Exception { get; set; }
+#endif
     }
 }