+++ src/Audit.SignalR/SignalrEventReconnect.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿#if ASP_NET
+using System.Collections.Generic;
 using Microsoft.AspNet.SignalR.Hubs;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
@@ -21,3 +22,4 @@ namespace Audit.SignalR
         public IHub HubReference { get; set; }
     }
 }
+#endif